05-monitoring/requirements.txt
prefect
tqdm
requests
joblib
pyarrow
psycopg2-binary
evidently==0.6.7
pandas
numpy
scikit-learn
jupyter
matplotlib